cpcbf/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpcbf"
version = "0.1.0"
description = "Cross-Platform Communication Benchmarking Framework"
requires-python = ">=3.10"
dependencies = [
    "paramiko>=3.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "scipy>=1.10",
    "jsonschema>=4.17",
    "psycopg2-binary>=2.9",
    "pyserial>=3.5",
]

[project.scripts]
cpcbf = "controller.cli:main"
cpcbf-ingest = "analysis.ingest:main"

[tool.setuptools.packages.find]
include = ["controller*", "analysis*"]
